Drag and drop basics in
Tiles can be moved between the panel and board. Tiles will push other tiles in the panel out of the way, so you can reorganize them. Board spaces will not let you put more than one tile in them. Tile updates to the next possible positions while you drag it, to give a preview, and is colored slightly different to make it obvious which on is moving (especially needed when dragging around panel)