Refactored drag and drop to use basic mouse event
HTML drag and drag was causing a couple issues for this project and it needed to be refactored. One is that d&d seems to override any css styling when it comes to the cursor. I'm overriding the basic behavior of d&d so I needed to change the cursor to reflect that. This doesn't seem to be possible, so I refactored it to use mouse events so that I could have control over it myself. The other issue this solves is that the d&d system wasn't catching the end of d&d if I hovered an area that was disallowed, even if I already manually moved the element, resulting in the tile remaining in the styling for dragging. I didn't chase that issue down much for HTML d&d because the cursor issue was more serious, but the mouse events just don't have this problem.