Added panels for the buttons and typed input
Added Button Panel that has buttons for game actions like Shuffle, Move, Restart, Return. Added Input Panel that allows you to type in the word you want and has buttons to enter that word and to cycle its location. None of these buttons functionality are implement, but the UI is in place.