Refactoring the game components, basic layout complete
Re-doing the Tiles as components that sit at the base of the game and move between Spaces. Made the same layout as before with the board and input panel now using Spaces. Rendering all the tiles in the bottom of the game overlapping right now, they will be getting move around separately.