Tiles animate smoothly between spaces
Tiles can be drag and dropped between spaces. When dragging a tile the potential spaces are outlined green. When the left mouse button is let go the tile gets its new space and animates to it. Also the tiles now gracefully load in, snapping to their initial positions and don't render until they are ready to do so.