Tiles handle window resize
Tiles were previously broken on window resize. They were not properly adjusting the variables that allow them to locate correctly. Changing the window size is not accounted for. There are still ways it might get skewed, but those seem to be rare, and moving any tile once now puts it back on track