Fixed solver error, added error tiles
Fixed an issue in the solver where it would consider some words that were made of tile placements not valid for the solution. If a word was valid (in the dictionary), it could be counted regardless of whether or not all of its tiles were placed correctly. So if a tile is part of two words, and one of them is in the dict and the other isnt, then neither of the words should count towards the solution. I added some logic to the solver to check for this before considering words for connectedness. Also added styling to Tile component to make this distinction obvious to players