Remade word list
Created the word list from a new base. The new base is much more permissive (which solves the issue of way too many words not being included before). So this requires a lot of processing to parse down all the not valid words. I remade the word-list-generator to use the new base list and to check against various other word lists I made to create the final word list and ts file. Made an api scraper that takes the filtered words and checks them against the free dictionary api  to create a list of possible exclusions.