Added mobile friendly tile selection movement
Added feature to tiles where you can click and release on the same tile to select it. Then if you click in an empty space, it will move the tile there. On desktop, hovered spaces get outlined when there is a selected tile, just like when there is a dragged tile. Dragging will cancel a selected tile. Great alternative to dragging the tile, especially on mobile where drag and drop can be clunky.