Spread game elements apart on mobile
On smaller screens, the game board is moved to the center with the tile and button panels moved to the bottom and title and score moved to the top. This makes it look like there is less dead space on mobile. On larger screens all of this content is instead centered to avoid giant gaps between content when the screen is taller.