Pulled letters and board size out of game
The letters of the puzzle were previously declared as literals in the Game component and the board size was declared in literals in several places. Created a single source of truth for both of those in the form of Game component props. Everything that has access to the Game Context can see them, and all the cases where width and height were assumed to be 10 and 14 are not referenced to that context. The board and the tile panel both adapt in size to match the board size and amount of letters.