Batching tile moves
Remade the moveTile method to moveTiles which now takes an array of movements in the form of the same tileID to new spaceID format. Edited the callers create an array of movements to give instead of calling move a bunch of times. React was probably batching this under the hood but this will be important when I add undo states so pressing "shuffle" doesn't cause 30 undo states