Select several tiles at once
Changed the feature where you select a tile and then click on a space to move it, so that you can select multiple tile at once. This way you can select the tiles as you plan out your word, then tap each location in order to place the word. Less mouse/thumb movement. Refactored the areas of the code that dealt with the old selection system