Return button returns selected first
If there are selected tiles in the board, and the user presses return, the selected tiles will return first. Pressing the return button again will return any unused tiles, as normal. The return button will not be disabled when there are no unused tiles, only if there are selected tiles on the board. Whenever the selection is of board tiles, the return button gets a yellow ring to match the selection styling, to indicate how what it will return.