Undo history
Added game state that logs where all the tiles are after each move. Updated moveTiles for this and added flag to not update state history (like for undoing). Used game context to expose history info to the button panel. Implemented undo and redo as moving the history index and moving the tiles to that state. Keeping track of the state with the best solution so far. The star button is implements, creates the next state as the best state so far.