Added puzzle generator
Made a system that generates tons of valid letter sets for the game. Constructs a complete board out of common words, and then extracts the letters in the puzzle. This way there is always a guaranteed set, and it will have letters that are realistic to make a solution out of. Also reorganized the scripts folder.