Added Posthog analytics
Integrated Posthog analytics into the project, including session replay. This meant adding two components, one that wraps the whole app as a provider and another that keeps track of page changes made with next router. Also updated the way the app is dockerized to work with the new env values.