Login & Navbar
Added authentication with next auth. Added to prisma schema as needed to work with next auth. Set up to wok with Google, Discord, and Github login. Updated layout to wrap app in session info. Added a navbar to have somewhere to interact with login, inserted on layout. Styled navbar responsively and added images for buttons