Daily page checks if completed
The page where you do the daily puzzle will know if you've already completed the puzzle, and show information about your solution it exists. Otherwise it loads the game as normal. This was done by changing the daily puzzle api route to also check is the session user has submitted a puzzle yet and if they have it return the data for it in addition to its normal return